chore: bazelignore the cpp-jsonnet subdirectory
Otherwise if the submodule is initialized Bazel will try to build
it when running commands like bazel test //...:all. But that doesn't
work because the cpp-jsonnet directory is not part of the same
bazel module.

Perhaps there is a better way to fix it but just ignoring the
cpp-jsonnet directory seems to be ok.